requirements.txt
# lanchain
python-dotenv
pytesseract
Pillow
pypdfium2
streamlit
multiprocessing
pandas
requests